stripped the app down to its bare bones
