+++ models/todo.js
@@ -2,9 +2,8 @@ var mongoose = require('mongoose'),
     Schema = mongoose.Schema;
 
 var TodoSchema = new Schema({
-  title: String,
-  description: String,
-  done: Boolean
+  // Add Schema details here
+
 });
 
 var Todo = mongoose.model('Todo', TodoSchema);
