+++ public/stylesheets/main.css
@@ -4,7 +4,3 @@ a, a:focus {
 a:hover {
   color: #aaa;
 }
-.done {
-  text-decoration: line-through;
-  opacity: .5;
-}
